requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
huggingface-hub==0.19.4
huggingsound==0.1.6
librosa==0.9.2
gunicorn==20.1.0
sentencepiec==0.2.0